Move snapshot to dedicated route, slash returns versions
